requirements.txt
Cython==0.29.21
progressbar2==3.53.1
torchaudio==0.7.2
git+https://github.com/facebookresearch/CPC_audio.git@1ef59879abaf424118e6a6bcd0a6aa56a6d60ccc
git+https://github.com/fairseq/fairseq.git@8d66c7c0b2afa612a974714248ab2daa156bce2e
git+https://github.com/spokenlanguage/platalea.git@zerospeech21-vg
